+++ NEWS.rst
@@ -79,8 +79,12 @@ sfs (:numref:`sfs`):
   * Removed the ability to work on multiple clips at once, as it was not used.
   * Can ingest a provided albedo map (of same size as the input DEM). Option:
     ``--input-albedo``.
-  * Removed the RPC approximation. Option ``--use-approx-camera-models`` should 
-    be used instead with ISIS cameras.
+  * Removed the RPC approximation logic. Use instead the option
+    ``--use-approx-camera-models`` with ISIS cameras.
+  * Removed the option ``--float-cameras``. It is more reliable to optimize
+    the cameras beforehand, in bundle adjustment. 
+  * Removed obsolete options ``--float-dem-at-boundary``,
+    ``--float-sun-position``, ``--coarse-levels``.
 
 orbit_plot (:numref:`orbit_plot`):
   * Added the options ``--use-rmse``, ``--output-file``.